Correcciones de varias cosas: En especifico de agregar contrato, incluye navegabilidad, es necesario terminar algunos metodos de cliente(DAO, SERVICE, VO, etc) para poder terminar de implementar completamente la parte de contrato, quedo iniciada la parte para mostrar en el combo box de contrato la lista de clientes.

+++ PSP/src/com/porcupine/psp/model/service/ContratoService.java
@@ -22,7 +22,7 @@ import javax.persistence.EntityNotFoundException;
  *
  * @author Feanor
  */
-public class ContratoService implements IService<ContratoVO, Integer> {
+public class ContratoService implements IService<ContratoVO, Short> {
     
     private static ContratoService instance;
     
@@ -66,7 +66,7 @@ public class ContratoService implements IService<ContratoVO, Integer> {
     }
 
     @Override
-    public ContratoVO find(Integer id) throws EntityNotFoundException, InsufficientPermissionsException {
+    public ContratoVO find(Short id) throws EntityNotFoundException, InsufficientPermissionsException {
         Contrato contrato = DAOFactory.getInstance().getContratoDAO().find(id);
         if (contrato != null) {
             return contrato.toVO();
@@ -108,7 +108,7 @@ public class ContratoService implements IService<ContratoVO, Integer> {
     }
 
     @Override
-    public void delete(Integer id) throws NonexistentEntityException, InsufficientPermissionsException {
+    public void delete(Short id) throws NonexistentEntityException, InsufficientPermissionsException {
         Contrato contrato = DAOFactory.getInstance().getContratoDAO().find(id);
         if (contrato != null) {
             DAOFactory.getInstance().getContratoDAO().delete(id);
